remaining 64bit fixes to build the main scalos binary.
git-svn-id: https://svn.aros.org/svn/aros/trunk@51949 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/scalos/main/ThumbnailCache.c
@@ -1306,7 +1306,7 @@ static void ThumbnailCacheGetEndOfLife(T_TIMEVAL *tv, ULONG LifetimeDays)
 	if (LifetimeDays)
 		tv->tv_secs += 24 * 60 * 60 * LifetimeDays;
 	else
-		tv->tv_secs = ULONG_MAX;	// forever
+		tv->tv_secs = INT_MAX;	// forever
 }
 
 //-----------------------------------------------------------------------
